own file for KeyFactItemBlock
